dev-aermod-agent-cleanup: moved location property to PointSource
