+++ JPS_VIRTUALSENSOR/AermodAgent/src/main/java/com/cmclinnovations/aermod/objects/Ship.java
@@ -1,10 +1,7 @@
 package com.cmclinnovations.aermod.objects;
 
-import org.postgis.Point;
-
 public class Ship extends PointSource {
     private String iri;
-    private Point location;
 
     public Ship(String iri) {
         this.iri = iri;
@@ -12,14 +9,6 @@ public class Ship extends PointSource {
 
     public String getIri() {
         return this.iri;
-    }
-
-    public void setLocation(Point location) {
-        this.location = location;
-    }
-
-    public Point getLocation() {
-        return this.location;
     }
 
     @Override
